+++ modules/globebrowsing/shaders/globalrenderer_vs.glsl
@@ -45,14 +45,14 @@ out vec3 positionCameraSpace;
 
 #if USE_ECLIPSE_SHADOWS
   out vec3 positionWorldSpace;
-  uniform dmat4 modelTransform;
+  uniform mat4 modelTransform;
 #endif // USE_ECLIPSE_SHADOWS
 
 #if SHADOW_MAPPING_ENABLED
   // ShadowMatrix is the matrix defined by:
   // textureCoordsMatrix * projectionMatrix * combinedViewMatrix * modelMatrix
   // where textureCoordsMatrix is just a scale and bias computation: [-1,1] to [0,1]
-  uniform dmat4 shadowMatrix;
+  uniform mat4 shadowMatrix;
   out vec4 shadowCoords;
 #endif // SHADOW_MAPPING_ENABLED
 
@@ -136,10 +136,10 @@ void main() {
   positionCameraSpace = vec3(modelViewTransform * vec4(pair.position, 1.0));
 
 #if USE_ECLIPSE_SHADOWS
-  positionWorldSpace = vec3(modelTransform * dvec4(pair.position, 1.0));
+  positionWorldSpace = vec3(modelTransform * vec4(pair.position, 1.0));
 #endif // USE_ECLIPSE_SHADOWS
 
 #if SHADOW_MAPPING_ENABLED
-  shadowCoords = vec4(shadowMatrix * dvec4(pair.position, 1.0));
+  shadowCoords = vec4(shadowMatrix * vec4(pair.position, 1.0));
 #endif // SHADOW_MAPPING_ENABLED
 }
